Extract nat events to separate package
