+++ nat/natevents/events_sender.go
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package traversal
+package natevents
 
 import (
 	log "github.com/cihub/seelog"
