+++ nat/natevents/events_sender_test.go
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package traversal
+package natevents
 
 import (
 	"errors"
