+++ nat/natevents/events_tracker.go
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package traversal
+package natevents
 
 import (
 	"time"
